requirements.txt
django
djangorestframework
django-filter
django-cors-headers
djangorestframework-simplejwt
joblib
pandas
numpy
scikit-learn
mysqlclient
drf-yasg